Remove use of oteltest in otelgin
