+++ instrumentation/github.com/gin-gonic/gin/otelgin/gintrace_test.go
@@ -18,169 +18,23 @@ package otelgin
 
 import (
 	"context"
-	"errors"
-	"html/template"
 	"net/http"
 	"net/http/httptest"
 	"testing"
 
 	"github.com/gin-gonic/gin"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 
 	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/codes"
-	"go.opentelemetry.io/otel/oteltest"
 	"go.opentelemetry.io/otel/propagation"
 	"go.opentelemetry.io/otel/trace"
 
 	b3prop "go.opentelemetry.io/contrib/propagators/b3"
-	"go.opentelemetry.io/otel/attribute"
 	oteltrace "go.opentelemetry.io/otel/trace"
 )
 
 func init() {
 	gin.SetMode(gin.ReleaseMode) // silence annoying log msgs
-}
-
-func TestChildSpanFromGlobalTracer(t *testing.T) {
-	otel.SetTracerProvider(oteltest.NewTracerProvider())
-
-	router := gin.New()
-	router.Use(Middleware("foobar"))
-	router.GET("/user/:id", func(c *gin.Context) {
-		span := oteltrace.SpanFromContext(c.Request.Context())
-		_, ok := span.(*oteltest.Span)
-		assert.True(t, ok)
-	})
-
-	r := httptest.NewRequest("GET", "/user/123", nil)
-	w := httptest.NewRecorder()
-
-	router.ServeHTTP(w, r)
-}
-
-func TestChildSpanFromCustomTracer(t *testing.T) {
-	provider := oteltest.NewTracerProvider()
-
-	router := gin.New()
-	router.Use(Middleware("foobar", WithTracerProvider(provider)))
-	router.GET("/user/:id", func(c *gin.Context) {
-		span := oteltrace.SpanFromContext(c.Request.Context())
-		_, ok := span.(*oteltest.Span)
-		assert.True(t, ok)
-	})
-
-	r := httptest.NewRequest("GET", "/user/123", nil)
-	w := httptest.NewRecorder()
-
-	router.ServeHTTP(w, r)
-}
-
-func TestTrace200(t *testing.T) {
-	sr := new(oteltest.SpanRecorder)
-	provider := oteltest.NewTracerProvider(oteltest.WithSpanRecorder(sr))
-
-	router := gin.New()
-	router.Use(Middleware("foobar", WithTracerProvider(provider)))
-	router.GET("/user/:id", func(c *gin.Context) {
-		span := oteltrace.SpanFromContext(c.Request.Context())
-		mspan, ok := span.(*oteltest.Span)
-		require.True(t, ok)
-		assert.Equal(t, attribute.StringValue("foobar"), mspan.Attributes()["http.server_name"])
-		id := c.Param("id")
-		_, _ = c.Writer.Write([]byte(id))
-	})
-
-	r := httptest.NewRequest("GET", "/user/123", nil)
-	w := httptest.NewRecorder()
-
-	// do and verify the request
-	router.ServeHTTP(w, r)
-	response := w.Result()
-	require.Equal(t, http.StatusOK, response.StatusCode)
-
-	// verify traces look good
-	spans := sr.Completed()
-	require.Len(t, spans, 1)
-	span := spans[0]
-	assert.Equal(t, "/user/:id", span.Name())
-	assert.Equal(t, oteltrace.SpanKindServer, span.SpanKind())
-	assert.Equal(t, attribute.StringValue("foobar"), span.Attributes()["http.server_name"])
-	assert.Equal(t, attribute.IntValue(http.StatusOK), span.Attributes()["http.status_code"])
-	assert.Equal(t, attribute.StringValue("GET"), span.Attributes()["http.method"])
-	assert.Equal(t, attribute.StringValue("/user/123"), span.Attributes()["http.target"])
-	assert.Equal(t, attribute.StringValue("/user/:id"), span.Attributes()["http.route"])
-}
-
-func TestError(t *testing.T) {
-	sr := new(oteltest.SpanRecorder)
-	provider := oteltest.NewTracerProvider(oteltest.WithSpanRecorder(sr))
-
-	// setup
-	router := gin.New()
-	router.Use(Middleware("foobar", WithTracerProvider(provider)))
-
-	// configure a handler that returns an error and 5xx status
-	// code
-	router.GET("/server_err", func(c *gin.Context) {
-		_ = c.AbortWithError(http.StatusInternalServerError, errors.New("oh no"))
-	})
-	r := httptest.NewRequest("GET", "/server_err", nil)
-	w := httptest.NewRecorder()
-	router.ServeHTTP(w, r)
-	response := w.Result()
-	assert.Equal(t, http.StatusInternalServerError, response.StatusCode)
-
-	// verify the errors and status are correct
-	spans := sr.Completed()
-	require.Len(t, spans, 1)
-	span := spans[0]
-	assert.Equal(t, "/server_err", span.Name())
-	assert.Equal(t, attribute.StringValue("foobar"), span.Attributes()["http.server_name"])
-	assert.Equal(t, attribute.IntValue(http.StatusInternalServerError), span.Attributes()["http.status_code"])
-	assert.Equal(t, attribute.StringValue("Error #01: oh no\n"), span.Attributes()["gin.errors"])
-	// server errors set the status
-	assert.Equal(t, codes.Error, span.StatusCode())
-}
-
-func TestHTML(t *testing.T) {
-	sr := new(oteltest.SpanRecorder)
-	provider := oteltest.NewTracerProvider(oteltest.WithSpanRecorder(sr))
-
-	// setup
-	router := gin.New()
-	router.Use(Middleware("foobar", WithTracerProvider(provider)))
-
-	// add a template
-	tmpl := template.Must(template.New("hello").Parse("hello {{.}}"))
-	router.SetHTMLTemplate(tmpl)
-
-	// a handler with an error and make the requests
-	router.GET("/hello", func(c *gin.Context) {
-		HTML(c, 200, "hello", "world")
-	})
-	r := httptest.NewRequest("GET", "/hello", nil)
-	w := httptest.NewRecorder()
-	router.ServeHTTP(w, r)
-	response := w.Result()
-	assert.Equal(t, http.StatusOK, response.StatusCode)
-	assert.Equal(t, "hello world", w.Body.String())
-
-	// verify the errors and status are correct
-	spans := sr.Completed()
-	require.Len(t, spans, 2)
-	var tspan *oteltest.Span
-	for _, s := range spans {
-		// we need to pick up the span we're searching for, as the
-		// order is not guaranteed within the buffer
-		if s.Name() == "gin.renderer.html" {
-			tspan = s
-			break
-		}
-	}
-	require.NotNil(t, tspan)
-	assert.Equal(t, attribute.StringValue("hello"), tspan.Attributes()["go.template"])
 }
 
 func TestGetSpanNotInstrumented(t *testing.T) {
