+++ instrumentation/github.com/gin-gonic/gin/otelgin/go.mod
@@ -13,6 +13,5 @@ require (
 	go.opentelemetry.io/contrib v0.22.0
 	go.opentelemetry.io/contrib/propagators v0.22.0
 	go.opentelemetry.io/otel v1.0.0-RC2
-	go.opentelemetry.io/otel/oteltest v1.0.0-RC2
 	go.opentelemetry.io/otel/trace v1.0.0-RC2
 )
