+++ instrumentation/github.com/gin-gonic/gin/otelgin/test/gintrace_test.go
@@ -1,0 +1,176 @@
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Based on https://github.com/DataDog/dd-trace-go/blob/8fb554ff7cf694267f9077ae35e27ce4689ed8b6/contrib/gin-gonic/gin/gintrace_test.go
+
+package test
+
+import (
+	"errors"
+	"html/template"
+	"net/http"
+	"net/http/httptest"
+	"testing"
+
+	"github.com/gin-gonic/gin"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
+	"go.opentelemetry.io/contrib/instrumentation/github.com/gin-gonic/gin/otelgin"
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/codes"
+	sdktrace "go.opentelemetry.io/otel/sdk/trace"
+	"go.opentelemetry.io/otel/sdk/trace/tracetest"
+
+	"go.opentelemetry.io/otel/attribute"
+	oteltrace "go.opentelemetry.io/otel/trace"
+)
+
+func init() {
+	gin.SetMode(gin.ReleaseMode) // silence annoying log msgs
+}
+
+func TestChildSpanFromGlobalTracer(t *testing.T) {
+	sr := tracetest.NewSpanRecorder()
+	otel.SetTracerProvider(sdktrace.NewTracerProvider(sdktrace.WithSpanProcessor(sr)))
+
+	router := gin.New()
+	router.Use(otelgin.Middleware("foobar"))
+	router.GET("/user/:id", func(c *gin.Context) {})
+
+	r := httptest.NewRequest("GET", "/user/123", nil)
+	w := httptest.NewRecorder()
+
+	router.ServeHTTP(w, r)
+	assert.Len(t, sr.Ended(), 1)
+}
+
+func TestChildSpanFromCustomTracer(t *testing.T) {
+	sr := tracetest.NewSpanRecorder()
+	provider := sdktrace.NewTracerProvider(sdktrace.WithSpanProcessor(sr))
+
+	router := gin.New()
+	router.Use(otelgin.Middleware("foobar", otelgin.WithTracerProvider(provider)))
+	router.GET("/user/:id", func(c *gin.Context) {})
+
+	r := httptest.NewRequest("GET", "/user/123", nil)
+	w := httptest.NewRecorder()
+
+	router.ServeHTTP(w, r)
+	assert.Len(t, sr.Ended(), 1)
+}
+
+func TestTrace200(t *testing.T) {
+	sr := tracetest.NewSpanRecorder()
+	provider := sdktrace.NewTracerProvider(sdktrace.WithSpanProcessor(sr))
+
+	router := gin.New()
+	router.Use(otelgin.Middleware("foobar", otelgin.WithTracerProvider(provider)))
+	router.GET("/user/:id", func(c *gin.Context) {
+		id := c.Param("id")
+		_, _ = c.Writer.Write([]byte(id))
+	})
+
+	r := httptest.NewRequest("GET", "/user/123", nil)
+	w := httptest.NewRecorder()
+
+	// do and verify the request
+	router.ServeHTTP(w, r)
+	response := w.Result()
+	require.Equal(t, http.StatusOK, response.StatusCode)
+
+	// verify traces look good
+	spans := sr.Ended()
+	require.Len(t, spans, 1)
+	span := spans[0]
+	assert.Equal(t, "/user/:id", span.Name())
+	assert.Equal(t, oteltrace.SpanKindServer, span.SpanKind())
+	attr := span.Attributes()
+	assert.Contains(t, attr, attribute.String("http.server_name", "foobar"))
+	assert.Contains(t, attr, attribute.Int("http.status_code", http.StatusOK))
+	assert.Contains(t, attr, attribute.String("http.method", "GET"))
+	assert.Contains(t, attr, attribute.String("http.target", "/user/123"))
+	assert.Contains(t, attr, attribute.String("http.route", "/user/:id"))
+}
+
+func TestError(t *testing.T) {
+	sr := tracetest.NewSpanRecorder()
+	provider := sdktrace.NewTracerProvider(sdktrace.WithSpanProcessor(sr))
+
+	// setup
+	router := gin.New()
+	router.Use(otelgin.Middleware("foobar", otelgin.WithTracerProvider(provider)))
+
+	// configure a handler that returns an error and 5xx status
+	// code
+	router.GET("/server_err", func(c *gin.Context) {
+		_ = c.AbortWithError(http.StatusInternalServerError, errors.New("oh no"))
+	})
+	r := httptest.NewRequest("GET", "/server_err", nil)
+	w := httptest.NewRecorder()
+	router.ServeHTTP(w, r)
+	response := w.Result()
+	assert.Equal(t, http.StatusInternalServerError, response.StatusCode)
+
+	// verify the errors and status are correct
+	spans := sr.Ended()
+	require.Len(t, spans, 1)
+	span := spans[0]
+	assert.Equal(t, "/server_err", span.Name())
+	attr := span.Attributes()
+	assert.Contains(t, attr, attribute.String("http.server_name", "foobar"))
+	assert.Contains(t, attr, attribute.Int("http.status_code", http.StatusInternalServerError))
+	assert.Contains(t, attr, attribute.String("gin.errors", "Error #01: oh no\n"))
+	// server errors set the status
+	assert.Equal(t, codes.Error, span.Status().Code)
+}
+
+func TestHTML(t *testing.T) {
+	sr := tracetest.NewSpanRecorder()
+	provider := sdktrace.NewTracerProvider(sdktrace.WithSpanProcessor(sr))
+
+	// setup
+	router := gin.New()
+	router.Use(otelgin.Middleware("foobar", otelgin.WithTracerProvider(provider)))
+
+	// add a template
+	tmpl := template.Must(template.New("hello").Parse("hello {{.}}"))
+	router.SetHTMLTemplate(tmpl)
+
+	// a handler with an error and make the requests
+	router.GET("/hello", func(c *gin.Context) {
+		otelgin.HTML(c, 200, "hello", "world")
+	})
+	r := httptest.NewRequest("GET", "/hello", nil)
+	w := httptest.NewRecorder()
+	router.ServeHTTP(w, r)
+	response := w.Result()
+	assert.Equal(t, http.StatusOK, response.StatusCode)
+	assert.Equal(t, "hello world", w.Body.String())
+
+	// verify the errors and status are correct
+	spans := sr.Ended()
+	require.Len(t, spans, 2)
+	var tspan sdktrace.ReadOnlySpan
+	for _, s := range spans {
+		// we need to pick up the span we're searching for, as the
+		// order is not guaranteed within the buffer
+		if s.Name() == "gin.renderer.html" {
+			tspan = s
+			break
+		}
+	}
+	require.NotNil(t, tspan)
+	assert.Contains(t, tspan.Attributes(), attribute.String("go.template", "hello"))
+}
