+++ instrumentation/github.com/gin-gonic/gin/otelgin/test/go.mod
@@ -1,0 +1,14 @@
+module go.opentelemetry.io/contrib/instrumentation/github.com/gin-gonic/gin/otelgin/test
+
+go 1.16
+
+require (
+	github.com/gin-gonic/gin v1.7.3
+	github.com/stretchr/testify v1.7.0
+	go.opentelemetry.io/contrib/instrumentation/github.com/gin-gonic/gin/otelgin v0.22.0
+	go.opentelemetry.io/otel v1.0.0-RC2.0.20210812161231-a8bb0bf89f3b
+	go.opentelemetry.io/otel/sdk v1.0.0-RC2.0.20210812161231-a8bb0bf89f3b
+	go.opentelemetry.io/otel/trace v1.0.0-RC2
+)
+
+replace go.opentelemetry.io/contrib/instrumentation/github.com/gin-gonic/gin/otelgin => ../
